crypto-bot/requirements.txt
discord.py==2.5.2
FastAPI==0.116.1
python-dotenv==1.1.0
docker==2.7.0
psycopg_pool
psycopg
uvicorn[standard]>=0.35.0